# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      description: >
        Validates the submitted email and password against the users table.
        On success, sets a signed, httpOnly session cookie and returns user info.
        Returns 401 on invalid credentials.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Login not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/heartbeat:
    post:
      operationId: authHeartbeat
      tags: [auth]
      summary: Record user activity heartbeat
      description: >
        Records that the authenticated user is active right now. Updates the
        current open session's lastActivityAt timestamp. If the open session has
        been idle for more than 5 minutes (now − lastActivityAt > 300 s) the
        server closes it (logoutAt = lastActivityAt, durationSeconds computed)
        and immediately opens a fresh one. Call every ~60 s while the user is
        active. Fire-and-forget — clients may ignore the response.
      responses:
        "200":
          description: Heartbeat recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required:
                  - ok
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change own password
      description: >
        Authenticated users can change their own password. Must provide the
        current password for verification. Clears the mustChangePassword flag.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
        "400":
          description: New password too short
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated or wrong current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AuthCredentials:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthStatus:
      type: object
      properties:
        authenticated:
          type: boolean
        email:
          type: string
        displayName:
          type: ["string", "null"]
        role:
          type: string
        mustChangePassword:
          type: boolean
      required:
        - authenticated
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword
...